Move validation result interfaces away...
