+++ pom.xml
@@ -5,7 +5,7 @@
 
     <groupId>fi.helsinki.cs.tmc</groupId>
     <artifactId>tmc-checkstyle-runner</artifactId>
-    <version>1.0.3-SNAPSHOT</version>
+    <version>2.0.0-SNAPSHOT</version>
     <packaging>jar</packaging>
 
     <description>Custom Checkstyle runner for TMC</description>
@@ -68,6 +68,13 @@
     </properties>
 
     <dependencies>
+
+        <!-- TMC Langs Abstraction -->
+        <dependency>
+            <groupId>fi.helsinki.cs.tmc</groupId>
+            <artifactId>tmc-langs-abstraction</artifactId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
 
         <!-- Simple Logging Facade -->
         <dependency>
