--- src/main/java/fi/helsinki/cs/tmc/stylerunner/validation/ValidationError.java
@@ -1,10 +1,0 @@
-package fi.helsinki.cs.tmc.stylerunner.validation;
-
-public interface ValidationError {
-
-    int getColumn();
-    int getLine();
-    String getMessage();
-    String getSourceName();
-
-}
