--- src/main/java/fi/helsinki/cs/tmc/stylerunner/validation/ValidationResult.java
@@ -1,12 +1,0 @@
-package fi.helsinki.cs.tmc.stylerunner.validation;
-
-import java.io.File;
-import java.util.List;
-import java.util.Map;
-
-public interface ValidationResult {
-
-    Strategy getStrategy();
-    Map<File, List<ValidationError>> getValidationErrors();
-
-}
